src/booksapi-grpc/pyproject.toml
[tool.poetry]
name = "booksapi-grpc"
version = "0.1.0"
description = ""
authors = ["Varvara <116665835+varvara007@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "booksapi_grpc"}]

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.54.2"
grpcio-tools = "^1.54.2"
grpcio-reflection = "^1.54.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
